saving progress refactoring TimedTextEditor as separate
